Begin versioning
I am staring into the depths of a bottomless NixOS rabbit hole filled with possibilities and I don't know where to start. This is either a horrible mistake or a great idea. We'll find out in a few years!
In the event my future self is reading this and trying to remember how it all went wrong, it's because the notion of exactly reproducing my system on multiple devices seemed very appealing. Remember trying to match KDE settings on your laptop?