Remove gpg patch
This only existed because I created the repo *before* adding gpg, so configuration.nix had uncommitted changes. To keep these in the initial commit, I stashed the changes, but then I realised it was difficult to remove an initial commit, so I deleted .git to start over. Of course, this also deleted the stash I just made, so I manually copied the result of an earlier git diff into this file to rescue the changes with a quick patch command. And then I forgot to delete it before running git add.
TLDR; I stupid lol