Enable Vim configuration
This *probably* works, but I haven't tested it because I've not gotten
around to importing the common modules on nix-on-droid, which by
extension means home-manager is also not active.
Sigh... I really really ought to stop procrastinating and sort that out someday, which also applies to a lot of things in my life at the moment.