Bugfix: Git config was being replaced instead of extended
lib.mkDefault seems to completely discard the set when a new one is introduced, which makes quite a lot of sense since it's for providing defaults if none are otherwise specified.
It just so happened the new set only had one value; the signingkey.
Removing mkDefault allows Nix to merge the sets, as intended.