Use existing Steam options
I have discovered there are options to open the firewall within nixpkgs's steam.nix. The issue with glXChooseVisual has also been recognised, so I don't need to reinvent the wheel.
Kind of a shame, since this is the very first module I wrote, but there's no reason for it to exist lol