Hack to show avatar in greeter
I've looked everywhere and it seems there's no existing way in
plasma-manager to add a profile picture, presumably because it's
confined to $HOME due to being imported inside home-manager.

Because the picture needs to be accessed *before* the user logs in, it
must be stored outside $HOME, specifically in
/var/lib/AccountsService/icons/$USER. I spent hours looking for an option
in NixOS which could create a symlink there properly, but there doesn't
seem to be one. You can do it for /etc/ with environment.etc.<file>, but
there is no equivalent for /var.

That's okay, I'll do it myself.