Discover NixOS host folders programatically
This saves me from having to maintain a list of hosts in flake.nix.
I'm happy I managed to figure this out by myself :)
Maybe in future it will be tidy to reorganise the logic into a
nixos/default.nix?

I'm taking a shortcut by not verifying if everything in nixos/hosts is
a directory, but I think that should be fine.